requirements.txt
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytest-flask==1.2.0
python-decouple==3.4
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1